+++ graphrag/index/flows/create_final_entities.py
@@ -1,0 +1,81 @@
+# Copyright (c) 2024 Microsoft Corporation.
+# Licensed under the MIT License
+
+"""All the steps to transform final entities."""
+
+import pandas as pd
+from datashaper import (
+    VerbCallbacks,
+)
+
+from graphrag.index.cache import PipelineCache
+from graphrag.index.verbs.graph.unpack import unpack_graph_df
+from graphrag.index.verbs.text.embed.text_embed import text_embed_df
+from graphrag.index.verbs.text.split import text_split_df
+
+
+async def create_final_entities(
+    entity_graph: pd.DataFrame,
+    callbacks: VerbCallbacks,
+    cache: PipelineCache,
+    name_text_embed: dict,
+    description_text_embed: dict,
+) -> pd.DataFrame:
+    """All the steps to transform final entities."""
+    # Process nodes
+    nodes = (
+        unpack_graph_df(entity_graph, callbacks, "clustered_graph", "nodes")
+        .rename(columns={"label": "name"})
+        .loc[
+            :,
+            [
+                "id",
+                "name",
+                "type",
+                "description",
+                "human_readable_id",
+                "graph_embedding",
+                "source_id",
+            ],
+        ]
+        .drop_duplicates(subset="id")
+    )
+
+    nodes = nodes.loc[nodes["name"].notna()]
+
+    # Split 'source_id' column into 'text_unit_ids'
+    nodes = text_split_df(
+        nodes, column="source_id", separator=",", to="text_unit_ids"
+    ).drop(columns=["source_id"])
+
+    # Embed name if not skipped
+    if name_text_embed:
+        nodes = await text_embed_df(
+            nodes,
+            callbacks,
+            cache,
+            column="name",
+            strategy=name_text_embed["strategy"],
+            to="name_embedding",
+            embedding_name="entity_name",
+        )
+
+    # Embed description if not skipped
+    if description_text_embed:
+        # Concatenate 'name' and 'description' and embed
+        nodes = await text_embed_df(
+            nodes.assign(name_description=nodes["name"] + ":" + nodes["description"]),
+            callbacks,
+            cache,
+            column="name_description",
+            strategy=description_text_embed["strategy"],
+            to="description_embedding",
+            embedding_name="entity_name_description",
+        )
+
+        # Drop rows with NaN 'description_embedding' if not using vector store
+        if not description_text_embed.get("strategy", {}).get("vector_store"):
+            nodes = nodes.loc[nodes["description_embedding"].notna()]
+        nodes.drop(columns="name_description", inplace=True)
+
+    return nodes
