+++ graphrag/index/workflows/v1/create_final_entities.py
@@ -31,8 +31,12 @@ def build_steps(
             "args": {
                 "skip_name_embedding": skip_name_embedding,
                 "skip_description_embedding": skip_description_embedding,
-                "name_text_embed": entity_name_embed_config,
-                "description_text_embed": entity_name_description_embed_config,
+                "name_text_embed": entity_name_embed_config
+                if not skip_name_embedding
+                else None,
+                "description_text_embed": entity_name_description_embed_config
+                if not skip_description_embedding
+                else None,
             },
             "input": {"source": "workflow:create_base_entity_graph"},
         },
